Modify settings loader to read text
The settings file currently recognizes:

data-path = path to base game data (e.g. C:\DAGGER\ARENA2 for DaggerXL)
width = display width (minimum: 320)
height = display height (minimum: 200)
fullscreen = true or 1 for fullscreen (not supported on Linux)
vsync = true or 1 for vsync (not not handled on Linux)
emnulate-low-res = true or 1 to emulate a 320x200 screen (stretched to fit the
                   display)

This is subject to change. Empty lines are ignored, there is no support for
comments.