Disable DaggerXL's world cache for now
It tries to write it to DaggerXL/WorldCache.dat, relative to the current
working directory. This isn't acceptable since the CWD can be anywhere, and we
don't want to leave it in random locations.

On Windows, it may need to go in %AppData%. On other systems, it should be
$XDG_CACHE_HOME (or $HOME/.cache). This is slightly complicated by needing to
ensure the appropriate sub-directories exist, or else it will fail to write
anyway. The cache path should likely be handled by the Engine or EngineSettings
classes, which would ensure the appropriate sub-directories exist when writing.